+++ packages/dotcom-ui-header/src/index.tsx
@@ -87,12 +87,10 @@ function Header(props: THeaderProps) {
 
 Header.defaultProps = defaultProps
 
-function LogoOnly(props: Pick<THeaderProps, 'variant'>) {
+function LogoOnly(props: Pick<THeaderProps, 'variant' | 'showLogoLink'>) {
   return (
     <HeaderWrapper {...props}>
-      <TopWrapper>
-        <TopColumnCenterNoLink />
-      </TopWrapper>
+      <TopWrapper>{props.showLogoLink ? <TopColumnCenter /> : <TopColumnCenterNoLink />}</TopWrapper>
     </HeaderWrapper>
   )
 }