+++ changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.6.0
+
+- JS build plugin: (_breaking change_) Removed `enableRequireDefault` option
+- Sass build plugin: (_breaking change_) Disabled the `webpackImporter` option for `sass-loader` and added an option to re-enable it
+- Handlebars package: Added a new `{{#renderReactComponent}}` helper to aid JSX and Handlebars interoperability
+- App context middleware: Added `pageKitVersion` property to app context data
+- Assets middleware: (_breaking change_) replaced by `dotcom-middleware-asset-loader` package
+- Navigation middleware: Added a new `getCurrentPath` option
+- Header UI component: Added a new `showLogoLink` option
+- (_breaking change_) Node v12 is now the minimum supported version
+- (_breaking change_) Removed the unused resource hints package
+- (_breaking change_) We recommend switching the `dom-loaded` package for `ready-state`
+
+Please refer to the [0.5.x to 0.6.x migration guide](https://github.com/Financial-Times/dotcom-page-kit/wiki/Upgrading-from-0.5.x-to-0.6.x) for more details.
+
 ## 0.5.8
 
 - Asset loader package: Fixed issue with concatenation of absolute URLs when used as the `publicPath`